requirements.txt
nltk
torch
tensorflow
mlogger
torchvision
setuptools
numpy
visdom
tqdm
PyYAML
psutil
pandas
sklearn
matplotlib
sps
